Struct i8080::memory::Memory

pub struct Memory(_);

A 64K memory.

Implementations

impl Memory

pub fn new() -> Self

Creates a 64K memory.

pub fn load_files<P: AsRef<Path>>(
&mut self,
paths: &[P],
start_address: u16
) -> Result<u16>

Loads ROM files located at paths into memory starting at start_address.

Errors

This function will return an error (Error::FileNotFound, Error::Io, or Error::TooLargeFile) if a ROM file in paths cannot be read successfully or the total size of the ROM files is too largo to be loaded into memory starting at start_address.

pub fn load_file<P: AsRef<Path>>(
&mut self,
path: P,
start_address: u16
) -> Result<u16>

Loads a ROM file located at path into memory starting at start_address.

Errors

This function will return an error (Error::FileNotFound, Error::Io, or Error::TooLargeFile) if the ROM file at path cannot be read successfully or its size is too largo to be loaded into memory starting at start_address.