Struct i8080::sys::i8080::I8080

pub struct I8080 {
    registers: Registers,
    flags: Flags,
    memory: Memory,
    cycles: u64,
    pub halted: bool,
    interrupt_flip_flop: bool,
    interrupt_op_code: Option<u8>,
    rx_devices: Vec<RxDevice>,
    tx_devices: Vec<TxDevice>,
    from_time: SystemTime,
    pub interactive: bool,
    pub current_state: String,
}

Fields
registers: Registersflags: Flagsmemory: Memorycycles: u64halted: boolinterrupt_flip_flop: boolinterrupt_op_code: Option<u8>rx_devices: Vec<RxDevice>tx_devices: Vec<TxDevice>from_time: SystemTimeinteractive: boolcurrent_state: String
Implementations

impl I8080

pub(crate) fn execute(&mut self, inst: u8)

pub fn dev_out(&self, device: usize, val: u8)

Writes a byte (usually A) to the specified out-device

If no such in-device exists, nothing is written

Arguments

val: Value to pass to the out-device

fn dev_in(&self, device: usize) -> u8

Read a byte from the specified in-device and writes it to A

If no such in-device exists, will read all ones, i.e. 0xff

pub fn halt(&mut self)

Halts the CPU at the end of the current cycle (if using i8080::run)

Though not true to the CPU, sends an end of transmission byte to all out-devices. This allows the program to end and is akin to a shutdown signal I suppose… kinda

fn add(&mut self, val: u8, carry: bool)

Adds a value to the A register

This function takes a carry argument. If false this functions acts like a [ADD A, x] instruction where as if using the real value of the carry register, this acts like [ADC A, x].

Arguments

val: Value to add

carry: Carry flag

fn sub(&mut self, val: u8, carry: bool)

Subtracts a value from the A register

This function takes a carry argument. If false this functions acts like a [SUB A, x] instruction where as if using the real value of the carry register, this acts like [SBB A, x].

Arguments

val: Value to subtract

carry: Carry flag

fn ana(&mut self, val: u8)

Performs a logical AND between a given value and A

Arguments

val: Value to be ANDed with A

fn xra(&mut self, val: u8)

Performs a logical XOR between a given value and A

Arguments

val: Value to be XORed with A

fn ora(&mut self, val: u8)

Performs a logical OR between a given value and A

Arguments

val: Value to be ORed with A

fn cmp(&mut self, val: u8)

Comapres a value against the A register

A CMP instruction is essentially a [SUB A, x] but the result is temporary and discarded. Here, for ease, I am replacing A in self.sub(...) and then restoring it.

Arguments

val: Value to compare

fn inr(&mut self, val: u8) -> u8

Increment a value, setting flags

This instruction does not set the carry flag

Arguments

val: Value to increment

fn dcr(&mut self, val: u8) -> u8

Decrement a value, setting flags

This instruction does not set the carry flag

Arguments

val: Value to decrement

fn jmp(&mut self, addr: u16, condition: bool)

Continues execution at a given address based on a condition

Arguments

addr: Address to which PC should be set

condition: The condition to test before jumping

fn push(&mut self, val: u16)

Push a value onto the stack

The order of the bytes should be swapped for a true i8080 implementation. However, for ease, I am not doing this.

Arguments

val: Value to push onto the stack

fn pop(&mut self) -> u16

Pop a value from the stack

The order of the bytes should be swapped for a true i8080 implementation. However, for ease, I am not doing this.

Arguments

val: Value to push onto the stack

fn push_psw(&mut self)

Push A and the flags register to the stack

[A, Flags] -> stack

fn pop_psw(&mut self)

Push A and the flags register to the stack

[A, Flags] -> stack

fn xthl(&mut self)

Exchange the value pointed to by the stack pointer with the value of the HL register

fn xchg(&mut self)

Exchange the value of HL with DE

fn call(&mut self, addr: u16, condition: Option<bool>)

fn ret(&mut self, condition: Option<bool>)

fn daa(&mut self)

Decimal Adjust Accumulator, represents the 8-bit value as two, 4-bit, decimal numbers

fn dad(&mut self, val: u16)

Double add, adds a word (a double) to HL

fn rlc(&mut self)

Rotate A left once, no carry

fn rrc(&mut self)

Rotate A right once, no carry

fn ral(&mut self)

Rotate A left once, with carry

fn rar(&mut self)

Rotate A right once, with carry

impl I8080

pub fn new(rx_devices: Vec<RxDevice>, tx_devices: Vec<TxDevice>) -> Self

fn sleep_for_hz(&mut self)

pub fn run(&mut self, emulate_clock_speed: bool)

pub fn cycle(&mut self)

pub fn issue_interrupt(&mut self, inst: u8)

pub fn load(&mut self, addr: u16, prog: Vec<u8>) -> usize

pub fn randomize(&mut self)

pub(crate) fn get_memory_slice(&self, addr: u16, len: u16) -> Vec<u8>

pub(crate) fn get_pc(&self) -> u16

fn pc_inst(&self) -> u8

Get the instruction at PC

fn pc_argb(&self) -> u8

Get the byte argument at PC+1

fn pc_argw(&self) -> u16

Get the word argument at PC+1

fn fmt_instruction(
&self,
inst: u8,
meta: OpMeta,
pc: u16,
is_interrupt: bool
) -> String

fn log_components(&self)

fn log_cycle(&self)

pub fn describe_system(&self) -> String

Example of the output as the format string is not particularly clear…:

┌─────────────────────────┐ ┌────────────┐ PC │ 0x0000 │ │ Sign: 0 │ ├─────────────────────────┤ ├────────────┤ SP │ 0xba5f │ │ Zero: 0 │ ├────────────┬────────────┤ ├────────────┤ B │ 0x63 (‘c’) │ 0x53 (‘S’) │ C │ Aux: 0 │ ├────────────┼────────────┤ ├────────────┤ D │ 0xcc (’ ’) │ 0x48 (‘H’) │ E │ Parity: 1 │ ├────────────┼────────────┤ ├────────────┤ H │ 0x74 (‘t’) │ 0x1a (’ ‘) │ L │ Carry: 0 │ ├────────────┼────────────┘ └────────────┘ A │ 0xee (’ ’) │ └────────────┘